Add product in inventory
have to work on add product and make changes in database

+++ order-service/src/main/java/com/flashdrop/orderService/dto/CouponRequest.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 public class CouponRequest {
 
     private String couponCode;
-    private int productPrice;
+    private double productPrice;
     private String productId;
     private int quantity;
